+++ psutil/tests/__init__.py
@@ -506,6 +506,7 @@ def chdir(dirname):
 
 
 def create_exe(outpath, c_code=None):
+    """Creates an executable file in the given location."""
     assert not os.path.exists(outpath), outpath
     if which("gcc"):
         if c_code is None:
@@ -526,6 +527,9 @@ def create_exe(outpath, c_code=None):
             safe_rmpath(f.name)
     else:
         # fallback - use python's executable
+        if c_code is not None:
+            raise ValueError(
+                "can't specify c_code arg as gcc is not installed")
         shutil.copyfile(sys.executable, outpath)
         if POSIX:
             st = os.stat(outpath)